Fix the ImplicitObjectProperties canonicalizer rule
