+++ src/canonicalizer/canonicalizer.cc
@@ -21,6 +21,7 @@
 #include "rules/if_without_then_else.h"
 #include "rules/implicit_array_lower_bound.h"
 #include "rules/implicit_object_lower_bound.h"
+#include "rules/implicit_object_properties.h"
 #include "rules/max_contains_without_contains.h"
 #include "rules/min_contains_without_contains.h"
 #include "rules/null_as_const.h"
@@ -46,6 +47,7 @@ sourcemeta::jsonbinpack::Canonicalizer::Canonicalizer(
   this->bundle.template add<IfWithoutThenElse>();
   this->bundle.template add<ImplicitArrayLowerBound>();
   this->bundle.template add<ImplicitObjectLowerBound>();
+  this->bundle.template add<ImplicitObjectProperties>();
   this->bundle.template add<MaxContainsWithoutContains>();
   this->bundle.template add<MinContainsWithoutContains>();
   this->bundle.template add<NullAsConst>();
